Rewrite the main block to use argparse
A good way to preserve the testing values during dev is to start with
writing the option parsing stuff using those values as the defaults.
That way you have the configurability right from the start, but you
can use sane defaults in the case that you test right off the bat.